+++ main.py
@@ -24,7 +24,11 @@ async def on_ready():
     await logchannel.send("STARTED")
     await bot.sync_commands()
     await logchannel.edit(content="")
-    
+    bot.load_extension("status")
+    bot.load_extension("info")
+    bot.load_extension("fortnitewins")
+    bot.load_extension("playtime")
+    bot.load_extension("housekeeping")
     print(f'Logged in as {bot.user} (ID: {bot.user.id})')
     print('-------------------------------------------------')
 online_list = []
@@ -38,211 +42,7 @@ for index, x in enumerate(uuid_list):
     last_online.append(int(time.time()))
 gamers = []
 current_time = int(time.time())
-@tasks.loop(seconds=3)
-async def status():
-    for index, uuid in enumerate(uuid_list):
-        parse_json_apidata_hypixel = hypixelapi(uuid,api_key)
-        channel = bot.get_channel(mainchannel)
-        logchannel = bot.get_channel(loggingchannel)
-        current_time = int(time.time())
-        if debug:
-            logging.basicConfig(
-                filename="logs.log",
-                format='%(asctime)s %(levelname)-8s %(message)s',
-                level=logging.DEBUG,
-                datefmt='%Y-%m-%d %H:%M:%S')
-        if not debug:
-            logging.basicConfig(
-                filename="logs.log",
-                format='%(asctime)s %(levelname)-8s %(message)s',
-                level=logging.WARNING,
-                datefmt='%Y-%m-%d %H:%M:%S')
-        if not parse_json_apidata_hypixel['success']:
-            break
-        try: 
-            online_status[index] = parse_json_apidata_hypixel['session']['online']
-        except Exception:
-            logging.warning("API ERROR")
-            await logchannel.send("API error perhaps (I want to die)" + str(parse_json_apidata_hypixel))
-            online_status[index] = "Questionable variable assignment to make api drop a straight nuclear shit in my bed causing the whole program to erupt"
-            pass
-        username = username_list[index]
-        if debug: 
-            print(online_status[index], username, online_list[index], online_list,last_online)
-            logging.debug(online_status[index], online_list[index], username, last_online[index], timestamper(current_time - last_online[index]))
-        if online_status[index]:
-            statusname = "ONLINE " + onlineemoji
-            online_status[index] = 'True'
-            online_time = ""
-        if not online_status[index]:
-            statusname = "OFFLINE " + offlineemoji
-            online_status[index] = 'False'
-            if uptime:
-                online_time = " They were online for: " + timestamper(current_time - last_online[index])
-            else: 
-                online_time = ""
-        if online_status[index] != online_list[index]:
-            await channel.send(modifier[index] + username + " has been " + statusname + " since <t:" + str(current_time) + ":R>" + online_time)
-            online_list[index] = online_status[index]
-            if online_status[index] == 'True':
-                gamers.append(username)
-                last_online[index] = current_time
-                print("writing", totaltime)
-                with open('totaltime.py', 'w+') as fp:
-                    fp.write("totaltime = [")
-                    for x in totaltime:
-                        fp.write("%s," %x)
-                    fp.write("]")
-                    fp.close()
-            elif online_status[index] == 'False':
-                gamers.remove(username)
-                timeplayed = current_time-last_online[index]
-                totaltime[index] = totaltime[index]+timeplayed
-                with open('totaltime.py', 'w+') as fp:
-                    fp.write("totaltime = [")
-                    for x in totaltime:
-                        fp.write("%s," %x)
-                    fp.write("]")
-                    fp.close()
-                last_online[index] = current_time
-        else:
-            pass
-        if len(gamers) > 1:
-            separator = ", "
-            await bot.change_presence(activity=discord.Game(name=separator.join(gamers) + " are online"))
-        elif len(gamers) == 1:
-            separator = ", "
-            await bot.change_presence(activity=discord.Game(name= separator.join(gamers) + " is online")) 
-        elif len(gamers) == 0:
-            await bot.change_presence(activity=discord.Game(name="No one is online"))
-        await asyncio.sleep(1)
-parse_fortnite_api = fortniteapi()
-wins = parse_fortnite_api['data']['stats']['all']['overall']['wins']
-@tasks.loop(seconds=10)
-async def fortnitewins():
-    global wins
-    parse_fortnite_api = fortniteapi()
-    try:
-        newwins = parse_fortnite_api['data']['stats']['all']['overall']['wins']
-    except Exception:
-        logging.error("API ERROR")
-    fnchannel = bot.get_channel(fortnitechannel)
-    try:
-        if newwins > wins:
-            await fnchannel.send("BIG W " + fortniteusername + " GOT AN EPIC VICTORY ROYALE <@&1058524479959076975> ")
-            wins = newwins
-    except Exception:
-        print("Fuck api ig")
-@bot.slash_command(description="Sends the bot's latency.")
-async def ping(ctx):
-    await ctx.respond(f"Pong! Latency is {int(bot.latency * 1000)}ms")
 
-@bot.slash_command(description='Total playtime for every account')
-async def stats(ctx):
-    embed = discord.Embed(title="Stats" ,
-                          description="Total playtime for each account",
-                          color=discord.Color.dark_purple())
-    for index, username in enumerate(username_list):
-        total_time = timestamper(totaltime[index])
-        current_time = int(time.time())
-        if online_status[index] == 'True':
-            statusEmoji = " :green_square:"
-            onlineorno = "**Online**\n"
-            lastorsince = "Online since: <t:"
-        elif online_status[index] == 'False':
-            statusEmoji = " :red_square:"
-            onlineorno = "**Offline**\n"
-            lastorsince = "Last online: <t:"
-        else:
-            statusEmoji = ":question:"
-            onlineorno = "**SOMETHING BROKE PING NOLAN!!!!**\n"
-            print(online_status)
-        embed.add_field(name=username + statusEmoji,
-                        value=lastorsince + str(last_online[index]) + ":R> \n" + onlineorno + "Total time online: " + total_time,
-                        inline=False)
-    embed.set_footer(text="Made by Noly")
-    await ctx.respond(embed=embed)
-
-@bot.slash_command(description="Tech support for the tech support loop")
-async def tech_support(ctx):
-    restoremyfaithinhumanity.start()
-    moyai.start()
-    await ctx.respond("Why did you even need to use this command smh (noly is shit at coding)")
-
-@tasks.loop(seconds=30)
-async def restoremyfaithinhumanity():
-    #seems redundant, might fix everything killing itself instantly
-    logchannel = bot.get_channel(loggingchannel)
-
-    if not status.is_running():
-        if not fortnitewins.is_running():
-            try:
-                fortnitewins.start()
-                status.start()
-            except Exception:
-                pass
-            await logchannel.send("Everything restarted, something catastrophic probably happened <@319574411579752459> \nThis can also just be the program starting")
-        if fortnitewins.is_running():
-            try:
-                status.start()
-            except Exception:
-                pass
-            await logchannel.send("Status restarted (Fuck api(probably))")
-    if not fortnitewins.is_running():
-        try:
-            fortnitewins.start()
-        except Exception:
-            pass
-        await logchannel.send("Fortnitewins restarted (it broke)")
-
-    else:
-        pass
-    await asyncio.sleep(30)
-@bot.slash_command(description="Get statuses and general stats of the bot")
-async def info(ctx):
-    if status.is_running():
-        statusStatus = "Running   :green_square:"
-    if not status.is_running():
-        statusStatus = "Not running   :red_square:"
-    if restoremyfaithinhumanity.is_running():
-        skillIssue = "Running   :green_square:"
-    if not restoremyfaithinhumanity.is_running():
-        skillIssue = "Not running   :red_square:"
-    if fortnitewins.is_running():
-        fortnitestatus = "Running   :green_square:"
-    if not fortnitewins.is_running():
-        fortnitestatus = "Not running   :red_square:"
-    if mayorchannel.is_running():
-        mayorstatus = "Running   :green_square:"
-    if not mayorchannel.is_running():
-        mayorstatus = "Not running   :red_square:"
-    embed = discord.Embed(title="Info",
-    color=discord.Color.dark_purple()
-    )
-    embed.add_field(name="Status",
-
-    value=statusStatus,
-    inline=False
-    )
-    embed.add_field(name="Task keeper",
-    value=skillIssue,
-    inline=False)
-    embed.add_field(name="Fortnite Win Tracker",
-    value=fortnitestatus,
-    inline=False)
-    embed.add_field(name="Mayorchannel",
-    value=mayorstatus,
-    inline=False)
-    await ctx.respond(embed=embed)
-
-@tasks.loop(hours=1)
-async def moyai():
-    #THIS DOES NOT WORK
-    userchnl = bot.get_channel(dmuser)
-    print(dmuser)
-    print(userchnl)
-    await userchnl.send(":moyai:")
-    
 
 
 @bot.slash_command(description="Start the mayor channel")
